restaurant directory 생성, menu를 restaurant에 넣기
