+++ src/main/java/com/happy/delivery/application/restaurant/MenuService.java
@@ -1,0 +1,12 @@
+package com.happy.delivery.application.restaurant;
+
+import com.happy.delivery.application.restaurant.result.MenuResult;
+import java.util.List;
+
+/**
+ * MenuService.
+ */
+public interface MenuService {
+
+  List<MenuResult> menuView(Long storeId);
+}
