+++ src/main/java/com/happy/delivery/application/restaurant/RestaurantSearchService.java
@@ -1,0 +1,17 @@
+package com.happy.delivery.application.restaurant;
+
+import com.happy.delivery.application.restaurant.command.RestaurantSearchCommand;
+import com.happy.delivery.application.restaurant.result.RestaurantCategoryResult;
+import com.happy.delivery.application.restaurant.result.RestaurantResult;
+import java.util.List;
+
+/**
+ * RestaurantSearchService.
+ */
+public interface RestaurantSearchService {
+
+  List<RestaurantCategoryResult> getCategories();
+
+  List<RestaurantResult> restaurantSearchByCategory(String category,
+      RestaurantSearchCommand restaurantSearchCommand);
+}
