+++ src/main/java/com/happy/delivery/application/restaurant/result/RestaurantCategoryResult.java
@@ -1,0 +1,46 @@
+package com.happy.delivery.application.restaurant.result;
+
+import com.happy.delivery.domain.restaurant.RestaurantCategory;
+
+/**
+ * RestaurantCategoryResult.
+ */
+public class RestaurantCategoryResult {
+  private int id;
+  private String name;
+
+  /**
+   * RestaurantCategoryResult constructor.
+   */
+  public RestaurantCategoryResult(int id, String name) {
+    this.id = id;
+    this.name = name;
+  }
+
+  public int getId() {
+    return id;
+  }
+
+  public String getName() {
+    return name;
+  }
+
+  @Override
+  public String toString() {
+    return "RestaurantCategoryResult{" +
+        "id=" + id +
+        ", name='" + name + '\'' +
+        '}';
+  }
+
+  /**
+   * RestaurantCategoryResult.
+   * RestaurantCategory -> RestaurantCategoryResult.
+   */
+  public static RestaurantCategoryResult fromRestaurantCategory(RestaurantCategory category) {
+    return new RestaurantCategoryResult(
+        category.getId(),
+        category.getName()
+    );
+  }
+}
