--- src/main/java/com/happy/delivery/domain/menu/repository/MenuRepository.java
@@ -1,11 +1,0 @@
-package com.happy.delivery.domain.menu.repository;
-
-import com.happy.delivery.domain.menu.Menu;
-import java.util.List;
-
-/**
- * MenuRepository.
- */
-public interface MenuRepository {
-  List<Menu> getAllByStoreId(Long storeId);
-}
