+++ src/main/java/com/happy/delivery/domain/restaurant/repository/MenuRepository.java
@@ -1,0 +1,11 @@
+package com.happy.delivery.domain.restaurant.repository;
+
+import com.happy.delivery.domain.restaurant.Menu;
+import java.util.List;
+
+/**
+ * MenuRepository.
+ */
+public interface MenuRepository {
+  List<Menu> getAllByStoreId(Long storeId);
+}
