+++ src/main/java/com/happy/delivery/domain/restaurant/repository/RestaurantSearchRepository.java
@@ -1,0 +1,16 @@
+package com.happy.delivery.domain.restaurant.repository;
+
+import com.happy.delivery.domain.restaurant.Restaurant;
+import com.happy.delivery.domain.restaurant.RestaurantCategory;
+import java.util.List;
+
+/**
+ * RestaurantSearchRepository.
+ */
+public interface RestaurantSearchRepository {
+
+  List<RestaurantCategory> getAllCategories();
+
+  List<Restaurant> getAllRestaurantsByCategory(String categoryId,
+      Double longitude, Double latitude);
+}
