--- src/main/java/com/happy/delivery/domain/user/repository/RestaurantSearchRepository.java
@@ -1,16 +1,0 @@
-package com.happy.delivery.domain.user.repository;
-
-import com.happy.delivery.domain.common.restaurant.Restaurant;
-import com.happy.delivery.domain.common.restaurant.RestaurantCategory;
-import java.util.List;
-
-/**
- * RestaurantSearchRepository.
- */
-public interface RestaurantSearchRepository {
-
-  List<RestaurantCategory> getAllCategories();
-
-  List<Restaurant> getAllRestaurantsByCategory(String categoryId,
-      Double longitude, Double latitude);
-}
