--- src/main/java/com/happy/delivery/infra/mybatis/MenuMapper.java
@@ -1,13 +1,0 @@
-package com.happy.delivery.infra.mybatis;
-
-import com.happy.delivery.domain.menu.Menu;
-import java.util.List;
-import org.apache.ibatis.annotations.Mapper;
-
-/**
- * MenuMapper.
- */
-@Mapper
-public interface MenuMapper {
-  List<Menu> getAllByStoreId(Long storeId);
-}
