+++ src/main/java/com/happy/delivery/infra/mybatis/restaurant/MenuMapper.java
@@ -1,0 +1,13 @@
+package com.happy.delivery.infra.mybatis.restaurant;
+
+import com.happy.delivery.domain.restaurant.Menu;
+import java.util.List;
+import org.apache.ibatis.annotations.Mapper;
+
+/**
+ * MenuMapper.
+ */
+@Mapper
+public interface MenuMapper {
+  List<Menu> getAllByStoreId(Long storeId);
+}
