--- src/main/java/com/happy/delivery/infra/mybatis/user/RestaurantSearchMapper.java
@@ -1,17 +1,0 @@
-package com.happy.delivery.infra.mybatis.user;
-
-import com.happy.delivery.domain.common.restaurant.Restaurant;
-import com.happy.delivery.domain.common.restaurant.RestaurantCategory;
-import java.util.List;
-import org.apache.ibatis.annotations.Mapper;
-
-/**
- * RestaurantSearchMapper.
- */
-@Mapper
-public interface RestaurantSearchMapper {
-
-  List<RestaurantCategory> getAllCategories();
-
-  List<Restaurant> getAllRestaurantsByCategory(String category, Double longitude, Double latitude);
-}
