--- src/main/java/com/happy/delivery/presentation/menu/response/MenuResponse.java
@@ -1,31 +1,0 @@
-package com.happy.delivery.presentation.menu.response;
-
-/**
- * MenuResponse.
- */
-public class MenuResponse {
-
-  private Long menuId;
-
-  private String menuName;
-
-  private String menuDetail;
-
-  private Long menuPrice;
-
-  public Long getMenuId() {
-    return menuId;
-  }
-
-  public String getMenuName() {
-    return menuName;
-  }
-
-  public String getMenuDetail() {
-    return menuDetail;
-  }
-
-  public Long getMenuPrice() {
-    return menuPrice;
-  }
-}
