+++ src/main/java/com/happy/delivery/presentation/restaurant/response/MenuResponse.java
@@ -1,0 +1,31 @@
+package com.happy.delivery.presentation.restaurant.response;
+
+/**
+ * MenuResponse.
+ */
+public class MenuResponse {
+
+  private Long menuId;
+
+  private String menuName;
+
+  private String menuDetail;
+
+  private Long menuPrice;
+
+  public Long getMenuId() {
+    return menuId;
+  }
+
+  public String getMenuName() {
+    return menuName;
+  }
+
+  public String getMenuDetail() {
+    return menuDetail;
+  }
+
+  public Long getMenuPrice() {
+    return menuPrice;
+  }
+}
