+++ src/main/resources/mybatis/mapper/RestaurantCategory.xml
@@ -3,16 +3,16 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.happy.delivery.infra.mybatis.user.RestaurantSearchMapper">
+<mapper namespace="com.happy.delivery.infra.mybatis.restaurant.RestaurantSearchMapper">
   <select id="getAllCategories"
-    resultType="com.happy.delivery.domain.common.restaurant.RestaurantCategory">
+    resultType="com.happy.delivery.domain.restaurant.RestaurantCategory">
     SELECT id, name
     FROM restaurant_categories
     ORDER BY id
   </select>
 
   <select id="getAllRestaurantsByCategory"
-    resultType="com.happy.delivery.domain.common.restaurant.Restaurant">
+    resultType="com.happy.delivery.domain.restaurant.Restaurant">
     SELECT id, name, category, addressDetail
     FROM restaurants
     where category = #{category}
