+++ src/main/resources/mybatis/mapper/menu.xml
@@ -3,8 +3,8 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.happy.delivery.infra.mybatis.MenuMapper">
-  <select id="getAllByStoreId" resultType="com.happy.delivery.domain.menu.Menu">
+<mapper namespace="com.happy.delivery.infra.mybatis.restaurant.MenuMapper">
+  <select id="getAllByStoreId" resultType="com.happy.delivery.domain.restaurant.Menu">
     SELECT *
     FROM menus
     WHERE storeSid = ${storeId}
